Factorisation du code dela partie transformation, et catégorisation des types de transformation
